cr: address CR comments
