+++ go.mod
@@ -4,8 +4,8 @@ require (
 	github.com/NYTimes/encoding-wrapper v0.2.0
 	github.com/NYTimes/gizmo v1.2.8
 	github.com/NYTimes/gziphandler v1.1.1
-	github.com/aws/aws-sdk-go-v2 v0.9.0
 	github.com/aws/aws-sdk-go v1.20.19
+	github.com/aws/aws-sdk-go-v2 v0.9.0
 	github.com/bitmovin/bitmovin-go v1.29.0
 	github.com/flavioribeiro/zencoder v0.0.0-20161215190743-745874544382
 	github.com/fsouza/ctxlogger v1.5.5
